+++ .travis.yml
@@ -2,7 +2,7 @@ language: node_js
 node_js:
   - "8"
 script: 
-  - "jest"
+  - "npm run test"
 deploy:
   provider: npm
   email: arojunior@gmail.com
