--- src/assets/index.css
@@ -1,6 +1,0 @@
-/*body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-}
-*/
