--- src/routes/About/index.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const About = () => (
-  <div>
-    Now you are in <code>src/routes/About/index.js</code>
-  </div>
-)
-
-export default About
