--- src/routes/About/routes.js
@@ -1,9 +1,0 @@
-export default {
-  path: 'about',
-
-  getComponent(nextState, cb) {
-    require.ensure([], require => {
-      cb(null, require('./index').default)
-    })
-  }
-}
