--- src/routes/Home/MiddleBody.js
@@ -1,95 +1,0 @@
-import React, {Component} from 'react'
-
-export default class MiddleBody extends Component {
-    render () {
-        return (
-        	<div className="col-sm-6">
-			<div className="panel panel-info">
-
-				<div className="panel-body">
-					<div className="media">
-						<a className="media-left" href="">
-							<img alt="" className="media-object img-rounded" src="http://placehold.it/64x64"/>
-						</a>
-						<div className="media-body">
-							<h4 className="media-heading">Media heading</h4>
-							<p>Dolorem aspernatur rerum, iure? Culpa iste aperiam sequi, fuga, quasi rerum, eum, quo natus tenetur officia placeat.</p>
-							<ul className="nav nav-pills nav-pills-custom">
-								<li><a href=""><span className="glyphicon glyphicon-share-alt"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-retweet"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-star"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-option-horizontal"></span></a></li>
-							</ul>
-						</div>
-					</div>
-
-					<div className="media">
-						<a className="media-left" href="">
-							<img alt="" className="media-object img-rounded" src="http://placehold.it/64x64"/>
-						</a>
-						<div className="media-body">
-							<h4 className="media-heading">Media heading</h4>
-							<p>Dolorem aspernatur rerum, iure? Culpa iste aperiam sequi, fuga, quasi rerum, eum, quo natus tenetur officia placeat.</p>
-							<ul className="nav nav-pills nav-pills-custom">
-								<li><a href=""><span className="glyphicon glyphicon-share-alt"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-retweet"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-star"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-option-horizontal"></span></a></li>
-							</ul>
-						</div>
-
-					</div>
-					<div className="media">
-						<a className="media-left" href="">
-							<img alt="" className="media-object img-rounded" src="http://placehold.it/64x64"/>
-						</a>
-						<div className="media-body">
-							<h4 className="media-heading">Media heading</h4>
-							<p>Dolorem aspernatur rerum, iure? Culpa iste aperiam sequi, fuga, quasi rerum, eum, quo natus tenetur officia placeat.</p>
-							<ul className="nav nav-pills nav-pills-custom">
-								<li><a href=""><span className="glyphicon glyphicon-share-alt"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-retweet"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-star"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-option-horizontal"></span></a></li>
-							</ul>
-						</div>
-
-					</div>
-					<div className="media">
-						<a className="media-left" href="">
-							<img alt="" className="media-object img-rounded" src="http://placehold.it/64x64" />
-						</a>
-						<div className="media-body">
-							<h4 className="media-heading">Media heading</h4>
-							<p>Dolorem aspernatur rerum, iure? Culpa iste aperiam sequi, fuga, quasi rerum, eum, quo natus tenetur officia placeat.</p>
-							<ul className="nav nav-pills nav-pills-custom">
-								<li><a href=""><span className="glyphicon glyphicon-share-alt"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-retweet"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-star"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-option-horizontal"></span></a></li>
-							</ul>
-						</div>
-
-					</div>
-					<div className="media">
-						<a className="media-left" href="">
-							<img alt="" className="media-object img-rounded" src="http://placehold.it/64x64" />
-						</a>
-						<div className="media-body">
-							<h4 className="media-heading">Media heading</h4>
-							<p>Dolorem aspernatur rerum, iure? Culpa iste aperiam sequi, fuga, quasi rerum, eum, quo natus tenetur officia placeat.</p>
-							<ul className="nav nav-pills nav-pills-custom">
-								<li><a href=""><span className="glyphicon glyphicon-share-alt"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-retweet"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-star"></span></a></li>
-								<li><a href=""><span className="glyphicon glyphicon-option-horizontal"></span></a></li>
-							</ul>
-						</div>
-					</div>
-
-				</div>
-			</div>			
-		</div>
-        	)
-    }
-}
