+++ src/routes/Home/index.js
@@ -1,25 +1,22 @@
-import React, {Component} from 'react'
-import Header from './Header.js'
-import LeftSide from './LeftSide.js'
-import MiddleBody from './MiddleBody.js'
-import RightSide from './RightSide.js'
+import React, { Component } from 'react'
+import Header from '../../components/Header'
+import LeftSide from './components/LeftSide'
+import MiddleBody from './components/MiddleBody'
+import RightSide from './components/RightSide'
 import '../../assets/App.css'
-
-
 export default class App extends Component {
-    render () {
-        return (
-        	<div>
-        		<Header />
-        		<div className="container">
-	        		<div className="row">
-	        			<LeftSide />
-		        		<MiddleBody />
-		        		<RightSide />
-	        		</div>
-        		</div>
-        		
-        	</div>
-        	)
-    }
+  render() {
+    return (
+      <div>
+        <Header />
+        <div className="container">
+          <div className="row">
+            <LeftSide />
+            <MiddleBody />
+            <RightSide />
+          </div>
+        </div>
+      </div>
+    )
+  }
 }
